Release v3.6; see CHANGELOG for details

+++ x2engine/protected/modules/actions/views/actions/_view.php
@@ -1,38 +1,24 @@
 <?php
-/*****************************************************************************************
- * X2CRM Open Source Edition is a customer relationship management program developed by
- * X2Engine, Inc. Copyright (C) 2011-2013 X2Engine Inc.
- * 
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Affero General Public License version 3 as published by the
- * Free Software Foundation with the addition of the following permission added
- * to Section 15 as permitted in Section 7(a): FOR ANY PART OF THE COVERED WORK
- * IN WHICH THE COPYRIGHT IS OWNED BY X2ENGINE, X2ENGINE DISCLAIMS THE WARRANTY
- * OF NON INFRINGEMENT OF THIRD PARTY RIGHTS.
- * 
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more
- * details.
- * 
- * You should have received a copy of the GNU Affero General Public License along with
- * this program; if not, see http://www.gnu.org/licenses or write to the Free
- * Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
- * 02110-1301 USA.
- * 
- * You can contact X2Engine, Inc. P.O. Box 66752, Scotts Valley,
- * California 95067, USA. or at email address contact@x2engine.com.
- * 
- * The interactive user interfaces in modified source and object code versions
- * of this program must display Appropriate Legal Notices, as required under
- * Section 5 of the GNU Affero General Public License version 3.
- * 
- * In accordance with Section 7(b) of the GNU Affero General Public License version 3,
- * these Appropriate Legal Notices must retain the display of the "Powered by
- * X2Engine" logo. If the display of the logo is not reasonably feasible for
- * technical reasons, the Appropriate Legal Notices must display the words
- * "Powered by X2Engine".
- *****************************************************************************************/
+/* * *******************************************************************************
+ * Copyright (C) 2011-2013 X2Engine Inc. All Rights Reserved.
+ *
+ * X2Engine Inc.
+ * P.O. Box 66752
+ * Scotts Valley, California 95067 USA
+ *
+ * Company website: http://www.x2engine.com
+ * Community and support website: http://www.x2community.com
+ *
+ * X2Engine Inc. grants you a perpetual, non-exclusive, non-transferable license
+ * to install and use this Software for your internal business purposes.
+ * You shall not modify, distribute, license or sublicense the Software.
+ * Title, ownership, and all intellectual property rights in the Software belong
+ * exclusively to X2Engine.
+ *
+ * THIS SOFTWARE IS PROVIDED "AS IS" AND WITHOUT WARRANTIES OF ANY KIND, EITHER
+ * EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION THE IMPLIED WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, TITLE, AND NON-INFRINGEMENT.
+ * ****************************************************************************** */
 
 Yii::app()->clientScript->registerScript('deleteActionJs', "
 function deleteAction(actionId) {
@@ -132,7 +118,9 @@ if(empty($data->type) || $data->type == 'weblead'){
 }elseif($data->type == 'note'){
     echo Formatter::formatCompleteDate($data->completeDate);
 }elseif($data->type == 'call'){
-    echo Yii::t('actions', 'Call:').' '.Formatter::formatCompleteDate($data->completeDate); //Yii::app()->dateFormatter->format(Yii::app()->locale->getDateFormat("medium"),$data->completeDate);
+    echo Yii::t('actions', 'Call:').' '.($data->completeDate == $data->dueDate
+            ? Formatter::formatCompleteDate($data->completeDate)
+            : Formatter::formatTimeInterval($data->dueDate,$data->completeDate,'{start}; {decHours} '.Yii::t('app','hours')));
 }elseif($data->type == 'event'){
     echo '<b>'.CHtml::link(Yii::t('calendar', 'Event').': ', '#', array('class' => 'action-frame-link', 'data-action-id' => $data->id));
     if($data->allDay){
@@ -145,21 +133,27 @@ if(empty($data->type) || $data->type == 'weblead'){
             echo ' - '.Formatter::formatLongDateTime($data->completeDate);
     }
     echo '</b>';
+}elseif($data->type == 'time'){
+    echo Formatter::formatTimeInterval($data->dueDate,$data->completeDate);
 }
 ?>
         <div class="buttons">
         <?php
         if(!Yii::app()->user->isGuest){
             if(empty($data->type) || $data->type == 'weblead'){
-                if($data->complete == 'Yes')
+                if($data->complete == 'Yes' && Yii::app()->user->checkAccess('ActionsUncomplete',array('assignedTo'=>$data->assignedTo)))
                     echo CHtml::link(CHtml::image($themeUrl.'/images/icons/Uncomplete.png'), '#', array('class' => 'uncomplete-button', 'title' => $data->id, 'data-action-id' => $data->id));
-                else{
+                elseif(Yii::app()->user->checkAccess('ActionsComplete',array('assignedTo'=>$data->assignedTo))){
                     echo CHtml::link(CHtml::image($themeUrl.'/images/icons/Complete.png'), '#', array('class' => 'complete-button', 'title' => $data->id, 'data-action-id' => $data->id));
                 }
             }
             if($data->type != 'workflow'){
-                echo $data->type != 'attachment' ? ' '.CHtml::link(CHtml::image($themeUrl.'/images/icons/Edit.png'), array('/actions/actions/update', 'id' => $data->id, 'redirect' => 1), array()).' ' : "";
-                echo ' '.CHtml::link(CHtml::image($themeUrl.'/images/icons/Delete_Activity.png'), '#', array('onclick' => 'deleteAction('.$data->id.'); return false'));
+                if(Yii::app()->user->checkAccess('ActionsUpdate',array('assignedTo'=>$data->assignedTo))){
+                    echo $data->type != 'attachment' ? ' '.CHtml::link(CHtml::image($themeUrl.'/images/icons/Edit.png'), array('/actions/actions/update', 'id' => $data->id, 'redirect' => 1), array()).' ' : "";
+                }
+                if(Yii::app()->user->checkAccess('ActionsDelete',array('assignedTo'=>$data->assignedTo))){
+                    echo ' '.CHtml::link(CHtml::image($themeUrl.'/images/icons/Delete_Activity.png'), '#', array('onclick' => 'deleteAction('.$data->id.'); return false'));
+                }
             }
         }
         ?>
@@ -235,7 +229,7 @@ else if($type == 'workflow'){
                 $userLink = empty($userLink) ? Yii::t('actions', 'Anyone') : $userLink;
                 echo Yii::t('actions', 'Assigned to {name}', array('{name}' => $userLink)).$relString;
             }
-        }else if($data->type == 'note' || $data->type == 'call' || $data->type == 'emailOpened'){
+        }else if(in_array($data->type,array('note','call','emailOpened','time'))){
             echo $data->completedBy == 'Guest' ? "Guest" : User::getUserLinks($data->completedBy).$relString;
             // echo ' '.Formatter::formatDate($data->completeDate);
         }else if($data->type == 'attachment' && $data->completedBy != 'Email'){
